Fixed session bug when reloading and added branch in raw material list

+++ src/main/java/com/cpi/is/dao/DispatchingDAO.java
@@ -6,7 +6,7 @@ import com.cpi.is.entity.DispatchingEntity;
 
 public interface DispatchingDAO {
 	
-	List<DispatchingEntity> getData(Integer branchId) throws Exception;
+	List<DispatchingEntity> getData(Long branchId) throws Exception;
 	String saveData(DispatchingEntity data) throws Exception;
 	String deleteData(DispatchingEntity data) throws Exception;
